Simplify support for multiple architectures
No longer bundle a set of architecture-dependent compatibility drivers in the repository.
Instead, download them from CVMFS at build time (as it is already done for the CUDA SDK),
and store them in the local "drivers" directory - without any architecture-specific structure.

The script, bin/ and drivers/ directories are anyway supposed to be installed in an
architecture-specific directory.